Exceptions and tests

+++ src/main/java/com/javaproject/demo/Repository/IEmployeeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 
 
-public interface IEmployeeRepository extends JpaRepository<Employee, Integer>{
+public interface IEmployeeRepository extends JpaRepository<Employee, Long>{
 
 }
